refactor(cli): carve host group into cli/groups/host.py
Move the host group + all subcommands (install/status/trace/list/migrate/
uninstall/guard) and host-lifecycle helpers out of the monolith. host
imports its cross-module deps (skill packaged-source, plugin bucket
bootstrap + _install_one, _shared _contract_guarantee); onboard now imports
_resolve_runtime_mode from the host module. All 5 command groups now live
in cli/groups/. Full suite: 2296 passed.
